move inputbox.tsx to /src/components
